requirements.txt
Flask
python-dotenv==1.0.0
gunicorn
Werkzeug
openai
jq
langchain
langgraph
langchain-openai
langchain-community
langchain-groq
langchain-google-genai
pathlib
requests
google.generativeai
geopy
pymongo==4.6.1
sentence-transformers
einops